move class to package
