+++ module-2/exercise/trellodemo.py
@@ -1,4 +1,4 @@
-""" Marvel Character Query tool
+""" Trello query tool, from the Marvel Character Query tool
 
 Notes:
     * The API contains extra wonk for authentication, so the authentication is
@@ -19,7 +19,9 @@ import json
 import os
 import requests
 import sys
-import urllib.parse as UP
+
+
+from TrelloSession import MarvelSession
 
 def derive_proxy_settings(environ):
     """ read the environment for proxy variables and generate key-value pairs for a
@@ -40,53 +42,6 @@ PROXIES = dict(derive_proxy_settings(os.environ))
 
 PUBLIC_KEY = '4e7701ebe8f8158383368664c6e029dc'
 PRIVATE_KEY = 'c7dc815ef8c8eb1c53ce8f075da6d5210f0d1cde'
-
-def convert_dict_to_querystring( dictionary ):
-    """
-        This test is here because at first I hand-rolled the functionality, then
-        later googled for a better way and refactored. The doctests are the contract
-        for the function, and still stand.
-
-        >>> convert_dict_to_querystring( {'a': 'b', 'c': 'd'} )
-        '?a=b&c=d'
-        >>> convert_dict_to_querystring( {'a': 'b', 'c': 'hello there'} )
-        '?a=b&c=hello+there'
-        >>> convert_dict_to_querystring( {'a': 'b', 'c': 'hello"there'} )
-        '?a=b&c=hello%22there'
-    """
-    return '?' + UP.urlencode(dictionary,doseq=False)
-
-class MarvelSession(object):
-    def __init__(self, root_url, public_key, private_key):
-        self._root_url = root_url
-        self._public_key = public_key
-        self._private_key = private_key
-        self._sequence=0
-
-    def request_url(self, path, query_dict):
-        """ Encapsulate the wonk for the Marvel API authentication.
-        query_dict contains the arguments for the query, which are merged
-        into the needed auth query items. """
-
-        self._sequence+=1
-        ts=self._sequence
-
-        hashable = str(ts) + self._private_key + self._public_key
-        hash_string = hashlib.md5( hashable.encode('UTF-8')).hexdigest()
-
-        query_items = {
-            'ts': str(ts),
-            'hash': hash_string,
-            'apikey': self._public_key
-        }
-
-        # merge in the
-        for k in query_dict:
-            query_items[k] = query_dict[k]
-
-        query_string = convert_dict_to_querystring(query_items)
-
-        return f'{self._root_url}{path}{query_string}'
 
 def retrieve_json(url):
     """ REST API helper for calls returning const results.
